+++ src/mass_driver/models/forge.py
@@ -1,5 +1,7 @@
 """Base definition of Forge, applying changes across git Repositories"""
 
+from enum import Enum
+
 from pydantic import BaseModel
 
 BranchName = str
@@ -32,3 +34,23 @@ class Forge(BaseModel):
 
         underscore_attrs_are_private = True
         """Ensure that _api is treated private"""
+
+
+class PROutcome(str, Enum):
+    """The category of result after using a Forge over a single repository"""
+
+    PR_CREATED = "PR_CREATED"
+    """The PR was created correctly"""
+    PR_FAILED = "PR_FAILED"
+    """The PR failed to be created"""
+
+
+class PRResult(BaseModel):
+    """The result of applying a patch on a repo"""
+
+    outcome: PROutcome
+    """The kind of result that PR creation had"""
+    pr_html_url: str | None = None
+    """The HTML URL of the PR that was generated, if any"""
+    details: str | None = None
+    """Details of the PR creation of this repo, if any"""