+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The project uses semantic versioning (see [semver](https://semver.org)).
 - New file type `Activity` combines `Migration` and `Forge`
 - New `Forge` named `dummy` for testing purposes
 - New `git_push_first` boolean param in Forge to disable git pushing.
+- New `ActivityOutcome` to capture the full result of a migration/forge sequence
 
 ### Changed
 - Replace commands `run-migration` + `run-forge` by new `run`, using the